Break ISoftwareCurvePoint and SoftwareCurve into their own files
