+++ LibreHardwareMonitorLib/LibreHardwareMonitorLib.csproj
@@ -40,6 +40,8 @@
     <Compile Include="Hardware\Controller\AquaComputer\MPS.cs" />
     <Compile Include="Hardware\Controller\Heatmaster\Heatmaster.cs" />
     <Compile Include="Hardware\Controller\Heatmaster\HeatmasterGroup.cs" />
+    <Compile Include="Hardware\ISoftwareCurvePoint.cs" />
+    <Compile Include="Hardware\SoftwareCurve.cs" />
     <Compile Include="Interop\Ftd2xx.cs" />
     <Compile Include="Hardware\Controller\TBalancer\TBalancer.cs" />
     <Compile Include="Hardware\Controller\TBalancer\TBalancerGroup.cs" />
